+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * feat: 升级springboot到2.7.18
 * feat: 升级相关类库版本
 * feat: starter-redis增加多数据源支持
+* feat: starter-redis redisson支持多数据源
 * fix: 修复starter-database自定义mapper地址后，无法扫描默认的SuperMapper
 * refactor: 删除多余代码、工具类
 * refactor: 命名修改